Updated Dev and Test
This commit adds dev container configurations to the repo (and backs out
text in the readme saying that it doesn't). When I started this project,
dev containers weren't really all too common, and it seemed like it
would be more work than it was worth to be prescriptive about using one.
Fast forward to today and it seems like _most_ devs use them now, so
including some bare bones configuration doesn't seem like an impediment
for contributors.

With that change, I also updated the test configs so that the database
hostname is set to the container name (instead of localhost).

This also removes some cruft in the dev/test deps.